% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleverStemmer.R
\name{CleverStemmer}
\alias{CleverStemmer}
\title{This function performs stemming on a string and converts it into a character vector}
\usage{
CleverStemmer(text)
}
\arguments{
\item{text}{Input a string}
}
\value{
a character vector that has been stemmed
}
\description{
This function performs stemming on a string and converts it into a character vector
}
\examples{
CleverStemmer("running buying feeding")
}
