% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleverStopwords.r
\name{CleverStopwords}
\alias{CleverStopwords}
\title{This function returns a comprehensive list of English stopwords that
allows adding more customized words.}
\usage{
CleverStopwords(words)
}
\arguments{
\item{words}{A list of strings that needs to be included in the stopwords.}
}
\value{
stopwords_all A list of strings that includes the complete set of stopwords.
}
\description{
This function returns a comprehensive list of English stopwords that
allows adding more customized words.
}
\examples{
words <- list("would", "aaa")
CleverStopwords(words)
}
