% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleverWordCloud.R
\name{CleverWordCloud}
\alias{CleverWordCloud}
\title{This function generates a meaningful word cloud with customized stopwords}
\usage{
CleverWordCloud(text, my_stopwords)
}
\arguments{
\item{text}{Input a character vector}

\item{my_stopwords}{a list of strings}
}
\value{
An image
}
\description{
This function generates a meaningful word cloud with customized stopwords
}
\examples{
\dontrun{
CleverWordCloud(c("cry", "foot", "maximum", "is"), list("is"))
}
}
