% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleverStemmer.R
\name{CleverStemmer}
\alias{CleverStemmer}
\title{This function performs stemming on a string or a character vector and converts it into a character vector}
\arguments{
\item{text}{Input a string or a character vector}
}
\value{
a character vector that has been stemmed
}
\description{
This function performs stemming on a string or a character vector and converts it into a character vector
}
\examples{
CleverStemmer("running buying feeding")
CleverStemmer(c("running", "buying", "feeding"))
}
