% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleverWordCloud.R
\name{CleverWordCloud}
\alias{CleverWordCloud}
\title{This function generates a visually appealing word cloud with customized shape and stopwords.}
\usage{
CleverWordCloud(text, stopwords = STOPWORDS, shape = CLOUD, max_words = 100)
}
\arguments{
\item{text}{Input an array of text / strings.}

\item{stopwords}{Input an set of strings.}

\item{shape}{Input a string that represents the path to an image.}

\item{max_words}{Input an integer to indicate the maximum number of words included in the word cloud}
}
\value{
An image
}
\description{
This function generates a visually appealing word cloud with customized shape and stopwords.
}
\examples{
CleverWordCloud(sample_text, max_words=200)
}
