% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleverWordCloud.R
\name{CleverWordCloud}
\alias{CleverWordCloud}
\title{This function generates a visually appealing word cloud with customized shape and stopwords.}
\usage{
CleverWordCloud(cleverlemstem_output)
}
\arguments{
\item{data}{DataFrame containing the words and their frequencies}
}
\value{
An image
}
\description{
This function generates a visually appealing word cloud with customized shape and stopwords.
}
\examples{
CleverWordCloud(c("cry", "foot", "maximum"))
}
